Logging Changes (#77)
* Logger seems to work, updated locks

* Replace logger in SmartNumberArbiterSpec

* Replace logger in PropertyInspectorSpec

* Replace logger in ParsesTestIdSpec

* Replace logger in RobustPcaRunnerSpec

* Replace logger in RobustPcaSpec

* Replace logger in ExampleSpec

* Replace logger in MigrationSpec

* Replace logger in DslSpec

* Replace logger in AbstractMeasurementCollector

* Replace logger in SmartNumberArbiter

* Replace logger in ZScoreArbiter

* Replace logger in RobustPcaArbiter

* Replace logger in PercentageDegradationArbiter

* Replace logger in WarpGuicer

* Replace logger in WarpPropertyLike

* Replace logger in ParsesTestId

* Replace logger in MeasurementExtension

* Replace logger in FutureUtils

* Replace logger in WarpStopwatch

* Replace logger in WarpRuntimeException

* Replace logger in WrappedApacheDistribution

* Replace logger in RobustPcaRunner

* Replace logger in TwoSampleRegressionTest

* Replace logger in Connection

* Replace logger in CorePersistenceAware

* Replace logger in PersistenceAware

* Replace logger in InfluxDBClient

* Replace logger in CanDropCoreSchema

* Replace logger in TestId

* Replace logger in HeapHistogram

* Replace logger in HeapHistogramFilter

* Replace logger in AbstractMeasurementCollectionController

* Replace logger in GatlingJUnitRunner

* Replace logger in WarpPropertyManager

* stashing work before lock regen

* update dependencies, left a TODO though

* add line to default logback

* Not progressing further, this is hairier than I'd like

* Followed this thread a bit, not very confident in the approach here

* whoops

* ignoring file policy

* manual logger config

* WarpLogUtils logger configuration rewrite with more test cases

* Changes to WriterConfig, adding fileWriter

* Account for failure to cast to LoggerContext

* Update locks

* tests pass

* WarpLogUtils code reuse cleanup - get context

* WarpLogUtils cleanup and docs

* Remove commented out dependency

* Remove dangling todo

* Doc update

* Remove tinylog dependencies + Java logging support

* Update scaladocs

* hugo docs regen script, check if hugo exists before rm

* Hugo docs update

* Code review change: alphabetize dependency versions

* Code review change: type annotations

+++ gh-pages-src/build-pages.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+if ! command -v hugo &> /dev/null
+then
+    echo "hugo could not be found, please install and rerun"
+    exit 1
+fi
+
 rm -rf ../docs
 hugo -D
 mv docs ../
